Rename the React Native project
