+++ ReactNative/microstates/android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.basicsetstateexample",
+    package = "com.microstatesreactnative",
 )
 
 android_resource(
     name = "res",
-    package = "com.basicsetstateexample",
+    package = "com.microstatesreactnative",
     res = "src/main/res",
 )
