+++ ReactNative/microstates/android/app/build.gradle
@@ -98,7 +98,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.basicsetstateexample"
+        applicationId "com.microstatesreactnative"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
