+++ ReactNative/microstates/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">BasicSetStateExample</string>
+    <string name="app_name">MicrostatesReactNative</string>
 </resources>
