+++ ReactNative/microstates/android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'BasicSetStateExample'
+rootProject.name = 'MicrostatesReactNative'
 
 include ':app'
