+++ ReactNative/microstates/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "BasicSetStateExample",
-  "displayName": "BasicSetStateExample"
+  "name": "MicrostatesReactNative",
+  "displayName": "MicrostatesReactNative"
 }
