+++ ReactNative/microstates/ios/MicrostatesReactNative.xcodeproj/xcshareddata/xcschemes/MicrostatesReactNative.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "BasicSetStateExample.app"
-               BlueprintName = "BasicSetStateExample"
-               ReferencedContainer = "container:BasicSetStateExample.xcodeproj">
+               BuildableName = "MicrostatesReactNative.app"
+               BlueprintName = "MicrostatesReactNative"
+               ReferencedContainer = "container:MicrostatesReactNative.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "BasicSetStateExampleTests.xctest"
-               BlueprintName = "BasicSetStateExampleTests"
-               ReferencedContainer = "container:BasicSetStateExample.xcodeproj">
+               BuildableName = "MicrostatesReactNativeTests.xctest"
+               BlueprintName = "MicrostatesReactNativeTests"
+               ReferencedContainer = "container:MicrostatesReactNative.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "BasicSetStateExampleTests.xctest"
-               BlueprintName = "BasicSetStateExampleTests"
-               ReferencedContainer = "container:BasicSetStateExample.xcodeproj">
+               BuildableName = "MicrostatesReactNativeTests.xctest"
+               BlueprintName = "MicrostatesReactNativeTests"
+               ReferencedContainer = "container:MicrostatesReactNative.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BasicSetStateExample.app"
-            BlueprintName = "BasicSetStateExample"
-            ReferencedContainer = "container:BasicSetStateExample.xcodeproj">
+            BuildableName = "MicrostatesReactNative.app"
+            BlueprintName = "MicrostatesReactNative"
+            ReferencedContainer = "container:MicrostatesReactNative.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BasicSetStateExample.app"
-            BlueprintName = "BasicSetStateExample"
-            ReferencedContainer = "container:BasicSetStateExample.xcodeproj">
+            BuildableName = "MicrostatesReactNative.app"
+            BlueprintName = "MicrostatesReactNative"
+            ReferencedContainer = "container:MicrostatesReactNative.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "BasicSetStateExample.app"
-            BlueprintName = "BasicSetStateExample"
-            ReferencedContainer = "container:BasicSetStateExample.xcodeproj">
+            BuildableName = "MicrostatesReactNative.app"
+            BlueprintName = "MicrostatesReactNative"
+            ReferencedContainer = "container:MicrostatesReactNative.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
