+++ ReactNative/microstates/ios/MicrostatesReactNative/Images.xcassets/AppIcon.appiconset/Contents.json
