+++ ReactNative/microstates/ios/MicrostatesReactNativeTests/Info.plist
